Created backupninja playbook and role.
Still not ready but this is a first impression of how it will look.
Playbook actually runs right now checking Debian version and
installing backupninja and duplicity from jessie even on wheeze
machines due to a wheezy-version bug of duplicity. Have only ran it
with -C so no changes will be done on hosts.